requirements.txt
flask
gunicorn
Pillow
pytesseract
Flask-Limiter
redis